requirements.txt
streamlit
llama-index-core
llama-index-embeddings-huggingface
llama-index-llms-groq
llama-index-readers-file
torch
transformers
huggingface-hub